third_party/blink/web_tests/external/wpt/css/selectors/invalidation/input-in-range-in-has-with-readonly-expected.txt

This is a testharness.js-based test.
[FAIL] :in-range in :has() invalidation when setting readonly
  assert_equals: ancestor should be black (readonly input is not in-range) expected "rgb(0, 0, 0)" but got "rgb(0, 128, 0)"
[FAIL] :in-range in :has() invalidation when removing readonly
  assert_equals: ancestor should be black (readonly input is not in-range) expected "rgb(0, 0, 0)" but got "rgb(0, 128, 0)"
[FAIL] :out-of-range in :has() invalidation when setting readonly
  assert_equals: ancestor should be black (readonly input is not out-of-range) expected "rgb(0, 0, 0)" but got "rgb(0, 0, 255)"
[FAIL] :out-of-range in :has() invalidation when removing readonly
  assert_equals: ancestor should be black (readonly input is not out-of-range) expected "rgb(0, 0, 0)" but got "rgb(0, 0, 255)"
Harness: the test ran to completion.
